requirements.txt
python-dotenv==1.0.0
airtable-python-wrapper==0.15.3
metaapi-cloud-metastats-sdk==3.3.1
metaapi-cloud-sdk==21.5.2
APScheduler=3.10